requirements_min.txt
# Core web framework
fastapi
uvicorn

# LangChain ecosystem
langchain
langchain-community
langchain-huggingface
langchain-groq
chroma-hnswlib
chromadb
langgraph

# LLM embeddings and text processing
tiktoken


# Data handling and environment configuration
pydantic
python-dotenv
numpy
pandas

# HTTP and web scraping (for document loaders)
requests


# Utilities
jinja2
